Workaround for mp4 & webm outputs if internal subs are selected
When working with local files an incompatible codec id <-> codec tag
error is caused when multiplexed subtitles are active and the file being
written is mp4. Most of these issues were resolved in mpv long ago, but
mostly for Matroska output and audio / video streams. As such this
covers an edge case, but avoids requiring the user to manually disable
subtitles when writing out to mp4.

WebM only supports WebVTT subs so this is a separate issue, but the
result is the same if .srt or other types of subs are embedded; external
subs do not appear to be an issue here and the file will write (sans
subtitles) if that's the case, but if the subs are internal the recorder
will reject the process.

mkv works universally for these cases and will apparently also write the
subs into the file, but the mp4 and webm formats have higher precedence
in our script for compatibility (with web standards etc., not supported
formats) so this work around makes it easier to work with local files.

Anyway, what this does is simply disable the subtitles before writing
and enable them again when the writing has finished.

Fixes #10