Use the track subtitle id field instead of passing it to the callback
The reason we check for pending is to avoid toggling subs for iterated
file writes: we want to re-enable them only at the end of the batch
rather than between each file. Problem is any subsequent file request
will overwrite the id and since subs were turned off then they won't
be reset. So, since we already track this id (for hostchange) it's easy
to just make use of it and just check an external state. Also, there's
no need to check the queue as the queue should be empty if the pending
list is.