Prevent writing hidden files on Unix-like systems
So first of all, the previous commit made specifying reserved characters
for the title forbidden as sanitizing the title is necessary after
property expansion; previously only the automatically set title was
subject to this, so you could in fact include these characters when
using `force_title` if you so desired. We could try to determine the OS
and carry out replacement only if Windows is detected by calling
`mp.get_property("platform")`, but in practice this appears to return
`nil, error` on some builds.

Since the reserved characters are also issue-prone on Unix-like systems
in the sense that they commonly require quoting or escaping on the
command line, it's not really a problem to play it safe and simply
replace the characters. This justification should've probably been
included in that commit message rather than this one, but since this is
where we would deal with it anyway I figured I'd explain why here.

Anyway, a `media-title` with a leading period was always possible, but
since we substitute using periods as well it makes it a bit more likely
and we never made sure to check for dotfiles or empty file names so this
commit practically makes sure the title is proper across all operating
systems.