Use an empty string instead of `"no"` for the `force_` option defaults
This changes the default setting for the following options:

* `force_extension`;
* `force_title`.

Although the behaviour hasn't changed (with the exception of being able
to specify a title of "no" once the deprecation period has passed) this
makes things easier to follow & reason about. Although seemingly
inconsistent, the `autoend` and `quit` defaults were not changed as
passing "no" there makes more sense semantically when setting them;
those options only accept a specific time format rather than an
arbitrary string and empty strings for the title and extension don't
make much sense in the context of the script. Also, `force_` checks
occur more frequently than the enabled automation comparisons
(which only occur once each at validation along with a disabled check at
the `autoquit` handler) so this does make the script easier to maintain.
Additionally, all automation options default to `no`.

In any case, `quit=no` appears to be more natural in contrast, while
`force_title=` compared to `force_title=no` is a better alternative when
you consider that the title being forced is / should be the value of the
string. Passing nothing for the value at the command line when opening a
stream is simpler anyway and the next commit will add an alternative way
of resetting the `force_` options globally through script-message
commands at runtime.

Also, since empty strings for the title are now disallowed (explicitly
enforced since the property expansion commit), if someone really wanted
to forego prepending a title (e.g. they want 1.mkv, 2.mkv, 3.mkv...
etc.) they could of course edit the script so that the title is not
concatenated and the hyphen is not prefixed to the increment, but
officially supporting this at the expense of unnecessary complexity is
probably not a great idea especially since the custom changes necessary
to achieve this are trivial.

The deprecation warning will only be emitted if any of these two options
are set to "no". This warning will be removed after a period of time.