Track the user set title, make expansion dynamic
Completes #12

This fixes b90a1f4 which was bugged in the sense that the expansion was
not dynamic: once the title was expanded and set at initialization it
would remain so and so titles based on dynamic properties (such as
`duration`) would remain fixed for any subsequent writes. This could
have easily been fixed by simply expanding `opts.force_title` instead of
`file.title`, but that doesn't cover the case where the user temporarily
sets the title at runtime through `script-message`.

Now, because expansion will cause the user set string to differ from the
actual title we need a way to know the title has been changed and we
need to store it so that it can be expanded at each write.

The solution here is to simply track a separate state and mirror it in
most cases while using an empty string as a distinguisher for state
checks for when there's a distinction between `opts.force_title` and
`file.forced_title`. This is of course more complex than just tracking
the option, but it is the simplest approach as long as we're going to
allow temporary title changes at runtime.

Also, now the title is properly reset if the `media-title` resets / the
stream reloads. Previously you could have forced the title, reverted,
then reloaded, and the title would not have reflected the forced title
set at the config file or the command line.