Properly reset `file.title`, expand title at beginning of `cache_write`
In practice resetting across title boundaries doesn't seem to be
strictly necessary, but it should probably be done since we do this with
the format for good reason and the update order isn't guaranteed,
meaning we want to avoid keeping a previous title around if another
stream is loaded; for manual cache writing this typically doesn't
matter, but under automatic writes we want to make sure the proper title
is reflected, so now the state is reset when the `media-title` is
missing.

As for expansion, this approach is simply more parsimonious since we
don't set `file.title` directly anymore if the title is custom. Since we
expand at write doing so at the very beginning avoids having to add an
extra condition and going through a redundant comparison.