Truncate long paths
Fixes #13

If we were simply dealing with ASCII we could use `string.sub` instead
to extract the substring, but since UTF-8 code points can span multiple
bytes our truncation scheme needs to be a little bit more involved.
In order to maintain compatibility with older versions of Lua and LuaJIT
we do this manually since `utf8.len` is only available starting with
Lua 5.3.