Resolve relative paths
Completes #13

This is essentially part 2 of the filename truncation implementation and
fixes the previous commit so that the length of a relative path is
measured properly.

Since Lua has a pretty minimal standard library and mpv doesn't seem to
offer a utility to convert relative paths to fully qualified ones this
needed to be done manually. At the moment I'm not aware of any trick to
extract this information in a surefire way across all platforms; maybe
you could open a command processor and read out `cd` commands or
something, but that would be kinda hacky and might not work universally.

That being said, this implementation is simple enough with the basic
tools that Lua does provide. There are likely some edge cases here and
there, but for the most part this approach should work and is definitely
an improvement to what we had before where path length wasn't even taken
into account.