Update the `lite` branch
This covers primarily the switch to an empty string for the `force_`
option defaults; the rest consists of simplifications.

I decided not to include the property expansion and truncation commits
as the goal of this version is simplicity and as such it should be as
bare bones as possible while retaining the core functionality. While it
could be argued that filename truncation covers basic use, our approach
there is somewhat involved and the problem doesn't affect all platforms
so I'll leave that at the master branch.

Long paths due to auto-generated titles is mostly an issue when the
`media-title` is some randomly generated string or reflects some long
URL and when that's the case it's probably desirable to change the file
title anyway and since this version still retains the ability to do so
through script-opts at runtime then it's not strictly necessary to go
through hoops to try and avoid the error.