`mode_switch` & `label_override`: use `string.format` for error printing
This covers a regression introduced in the previous two commits; for
script-message commands with no arguments the first parameter inside of
these functions points to `nil` instead of being interpreted as an empty
string and so concatenation would fail. The setting of default values is
otherwise not necessary as explained in the previous commit.