Cover formatted error handling with older versions of Lua
Introducing quite a bit of complexity just to print an error message,
but this approach has the advantage of being robust and universal; the
catch branch isn't hit unless mpv is built with Lua 5.1 and the script
runs into one our self-defined errors and even then it's not really
expensive. The drawback is the bloat this adds just to cover that case,
but it makes the rest of the code easier to maintain going forward so in
total it's more parsimonious.

The reason that function exists in the first place is so that it can be
reused, making the main code much more readable. Concatenating requires
either carefully guarding with defaults or sprinkling `tostring`
everywhere; it's also longer and more obtuse for these specific kinds of
formatted messages.

Note: an alternative approach would be to check the `_VERSION` string as
well as the existence of the `jit` table instead of utilizing a
protected call. So branching with:

if _VERSION == "Lua 5.1" and not jit then
    (pack)
    (cast to strings)
    (unpack)
    (call)
else
    call(...)
end

But this is probably not as safe as the current approach.